Added filter to ignore "viz_bngl" directory in viz file directories.
The previous glob call included all files in the viz directory as viz
data files. However, now that we've added a new subdirectory for BNGL
viz data, the glob call also returns that directory ("viz_bngl"). This
commit adds a simple filter to remove any files ending with an os path
separator followed by the string "viz_bngl".

This is a crude filter, but it works for now. Without this filter,
CellBlender will report an error and show no molecules whenever that
directory is attempted to be displayed.