Fixed broken reactions for non-BNGL models
Commit 894e8103a0d1ad85e7affe743886ec0e6f8c3c50 on September 19th, 2018
introduced a bug in CellBlender's non-BNGL reactions. The bug was the
removal of forced added spacing around the "+" sign in reactions.

The BioNetGen language uses the "+" sign in situations where spaces
around the "+" (such as " + ") change the meaning (or are illegal).

Earlier versions of CellBlender (before Sept 19th, 2018) automatically
added spaces around any "+" sign in a reaction string. This was breaking
the BioNetGen reactions used with MCellR in "BioNetGen Language Mode",
so the spaces were removed.

This commit only removes the spaces in BioNetGen language mode.