Added an operator to convert strings from NAUTY format to BNGL.
This commit adds two operators, but only one of them is implemented:

  mol.to_bngl
  mol.to_nauty

The "to_bngl" operator is implemented and it does the conversion between
an MCellR "nauty" string and BNGL. The input and output are through
string properties in the MolMaker panel to support cutting and pasting
between other applications (primarily MCellR's output text).

The parsing of the string is relatively straightforward since the nauty
strings are machine-generated and machine consumed. The resulting BNGL
from this simple algorithm doesn't attempt to arrange the bonds in any
logical manner. They simply reflect the order implied from the nauty
string. The code also doesn't add the compartments to the molecules.
It's not clear if this would be useful, but the compartments are parsed,
but they're just not added to the resulting BNGL.

At this time, there isn't a need to process these in CellBlender. The
current need is within MCell (in C). But Python seemed like a good way
to get started, and it made sense to integrate it into CellBlender to
both track the Python code (in Git) and make it available as a utility.