Added special debug support for dropping to console and comments.
CellBlender has had a "debug_level" option for a while, and it usually
causes more output to be printed to the console. While developing the
new MolViz code, it was also useful to drop into a Python console during
parts of the code. This commit increased the range of debug levels from
100 to 101. Level 101 will enable the console is some sections of code.
Because this can be a problem if Blender is run without a console, this
option will show the red "alert" whenever the debug_level is above 100.