Added help for new "Show Proxy" option in MCell and default to show.
The new version of MCell will not default to generating proxy molecules
in MCellR mode. However, if they are desired, there's a command line
parameter (-z 0x08) to force output of all proxy molecules. This change
adds that to the help menu for the command line options.

Additionally, the previous versions of CellBlender automatically hid all
of the proxy molecules on every frame load. There was a global option in
the Preferences panel to over-ride that behavior and show them. But
since they will generally NOT be appearing in future viz files, this
version defaults to showing them if they do appear. They can always be
hidden in the outliner panel if they're not desired.