Moved the molecule label initialization into cellblender/__init__.py
As mentioned in the previous commit, the "register" functions in each of
our modules appear to NOT be called. So the placement of the molecule
label initialization code in "cellblender_molecules.py" was not working.

This commit moves that code into CellBlender's main initialization file
("cellblender/__init__.py") where it appears to be working.

This raises the question as to whether any of our "register/unregister"
functions are needed. This commit has commented out the reg/unreg
functions in cellblender_molecules.py. Maybe this should be done in
other places as well.