Attempt to find "None" error when zoomed in "User Perspective" mode
A "None" error is reported when zoomed in and in User Perspective mode.
This commit attempts to isolate the variable that is "None" in that
situation. However, none of the "print" statements appear to be
triggered, and the error still happens. This makes it appear that
something down inside one of the variables may have a None value.

For now, don't zoom in when in User Perspective mode with labels.