Added code to display labels when found in a "molecule_labels" object.
This version of the molecule label code expects custom code to create a
"molecule_labels" object which contains the labels. This was created as
an object that could hold a mesh in the expectation that the label
locations would be stored in the mesh. However, it was found to be
easier to store the labels and index values as Blender ID properties
rather than as coordinates. If this turns out to be acceptable, then
future versions might store the label coordinates as ID properties in
the existing MCell property groups rather than creating a new Blender
object ("molecule_labels") to store that data.