Pulled some of the dictionary references out of inner loop for speed.
The "draw_labels_callback" is called to draw the labels on every frame.
It uses several arrays which are stored in ID properties in Blender. The
previous version looked up the array with its ID property name for every
pass of the loop. This minor change looks up the arrays via ID property
name outside of the loop so the data can be pulled from arrays without
any lookup.