Consolidated label display in mol_viz and handled points "behind the lens".
The previous version created a molecule labels object to hold the 3D points
of the labels in a mesh. Since the points are being processed through Python,
it's not clear that there was any advantage to having them in a mesh. This
commit collects the points in "ID Properties" under the "mol_viz" group.
That eliminates the additional object which might have been confusing.

Also, previous versions produced a lot of error messages when zoomed in
while in "Perspective Mode". This was because the location_3d_to_region_2d
function was returning "None" when the points were "behind the lens". That
case is now checked in the updated "draw_labels_callback" function. Thanks
to Tom for finding that problem.