Quick patch for GitHub issue #109 ("Plotting issue on cluster").
This patch adds an "Ignore Start Time" flag to the plotting screen. The
code normally checks that the start_time.txt exists in the proj_files
directory. However, when running multiple simulations on a cluster, this
method may not be working properly. This patch allows the user to
disable this checking to allow plotting even when the start_time.txt
file isn't consistent with the plot files.

Note that the start_time.txt file was intended as a way to keep users
from accidentally plotting stale data files from previous runs along
with current plotting files from current runs. When the user ignores
the start_time.txt file, they should ensure that all of their runs have
completed properly.