Windows: Modified "makedirs_exist_ok" functions to support Windows
Windows paths start with a drive letter (such as "c:") which was not
handled in the original version. This current fix just checks for a
colon (":") in the first section after splitting by os.sep. If found,
then it's treated like a drive and becomes the root of the path.

Note that much of this functionality is included in Python's new
"pathlib" module (since Python 3.4). At some point it might be good
to update all of our path handling code to use that new module.