Windows: Modified Java Plotter to work in Windows
In Windows, the typical "java" executable is "java.exe" and not "java".
This commit checks the os.name to determine what name to find.

This version also uses the list version of Python's popen to better
handle paths with embedded spaces. The previous version generated one
string for the entire command.

This change to the Java Plotter was prompted by the desire to test the
Windows version of MCell before it is bundled into Blender. Some of the
Examples (like variable rate constants) are best tested with a plot, and
this change provides a quick means to do so.