Added code to support quoted strings as arguments to run_wrapper
The Queue Runner communicates commands to run_wrapper via the stdin
stream to run_wrapper. This presents a problem when using spaces as
delimiters in situations where spaces should not be delimiters (as
when file names or paths contain embedded spaces).

There are a number of ways to handle this situation, and simply adding
quotes to all strings was chosen as a reasonable approach. The JSON
encoding for strings was chosen to handle embedded quotes and escapes
(backslashes) as follows:

  Encode a quote (") with backslash quote (\")
  Encode a backslash with backslash backslash (\\)

With this system, the "run_wrapper" can assume that if the strings it
reads from stdin are enclosed in quotes, then this quoting system is
being used. If, however, the strings it reads from stdin are not
enclosed in quotes, then the old non-quoted system is being used.

This commit adds untested support for the quoted string encoding to
run_wrapper, but cellblender_simulation is still sending unquoted
strings. The transition to use quoted strings will be done in another
commit.