Saved a copy of changes to "run_wrapper.py" made on Windows.
For a number of reasons, the Windows version ended up requiring
different code than the Linux version. The biggest problem is the
Windows use of backslashes for path separators. Unfortunately, that's
also the "escape" character for special characters in JSON strings
and in Python.

The changes in this commit were made in a Windows VM and were simply
copied to the end of "run_wrapper.py" as a triple-quoted string. This
Windows code still needs to be integrated into the Linux code so that
the same code will work on both platforms (possibly by checking the
platform if needed).