Modified run_wrapper.py to choose between Windows and Posix mode.
The previous version contained the Windows code in a comment, but it was
not integrated into the actual code. This version checks the platform to
decide which versions of the code to run.

Note that this implementation pretty much duplicates all of the (few)
functions in run_wrapper.py. A more careful coding could integrate the
two versions more closely.