Added logic to force MCellR mode when BioNetGen Language mode is set.
The previous version only used the "BioNetGen Language" mode flag to
control how the molecule panel was drawn (with or without components).

That previous version ALWAYS auto-detected the MCellR mode using the
"requires_mcellr" function in "mdl/data_model_to_mdl.py". That function
looked for a "bngl_component_list" in the molecule definitions to make
its determination.

However, there is now a need to run models in MCellR mode that have no
components on any of their molecules. This could be done by either
looking for other "clues" or with a manual switch. Since the BioNetGen
Language mode flag already exists, that seems like the logical choice.

As a result, this commit should run a model in MCellR mode if either the
BioNetGen Language mode flag is set OR if the "bngl_component_list" is
not empty for any molecule in the model.

Note that there might be other approaches to detecting BioNetGen
Language mode in the model. For example, it's highly likely (maybe
mandatory) that BioNetGen Language models must use an "@" sign in their
release definitions. That could be used as another way to detect a
BioNetGen Language model. That has not been implemented in this commit.