Fixed the handling of parameter lists in sim_runner_queue.py
The code had been failing at: char_stream = '"' + a + '"'

This happened when some of the arguments were lists instead of strings.

This fix uses a recursive "flatten_list" to pull the elements out of
nested lists and put them at the top level list. It's not clear how or
why this is happening, but this fixed the problem.