Added support for newer compartment syntax not requiring ":"
The BNGL importer used for CBNGL (bng/bngl_to_data_model.py) had used
the ":" in the previous syntax as part of its compartment detection
logic. However, the ":" is no longer required with the M()@C syntax.

The symptom of this problem was missing compartments in release sites
when the new syntax was used. Note that there may be other changes
required to support the new syntax.