Improved the LR CBNGL formatting using CellBlender's Scripting Tools
The best way to make a script for an example is using the Data Model
Scripting panel with the various "Include" buttons checked and with the
section set to "ALL". Then the "Copy" button will produce a clipboard
copy of the data model in Python format. This is the easiest format to
copy into a python example file.