Added button to find Blender's Python in Windows (Pure Python Sim)
This commit adds a button titled "Blender's Python" which will set the
"Python Command" field to the path where Blender's Python exists on
Windows.

This has not been needed when running in Linux which typically contains
a valid "python" command in the user's path. However, on Windows, this
is much less likely. While the user could navigate to find Blender's own
Python in such cases, it's not very easy. So the os.__file__ value is
now used to deduce the location of the Python interpreter that Blender
is currently using. In Windows, the value of os.__file__ is typically:

  C:\Program Files\Blender Foundation\Blender\2.xx\python\lib\os.py

The actual interpreter is then expected to be at:

  C:\Program Files\Blender Foundation\Blender\2.xx\python\bin\python.exe

This has been tested in a Windows 10 Virtual Machine.

Note that the typical name for Blender's python in Linux has an
additional suffix (such as "python3.5m"). This is not currently handled
by this current code.