Added sys.platform detection to handle Blender's Python path differences
As mentioned in the earlier commit, Blender uses different paths on
Windows and Linux. This commit detects the platform and attempts to set
the proper Python path and command. Note that Blender's Linux version
uses "python3.5m" rather than simply "python" as the command. A future
version might do a better job of detecting the actual name as well.
However, this is generally not needed on Linux platforms anyway, since
they generally have a default "python" which will also work with our
Pure Python simulation.