Rolled the test suite main test back to an earlier commit
The last few commits of the test suite (early 2018) added a lot of code
to support automatic tutorial generation. That approach was never
completed. While that code may be worth adding again at some point, it's
currently more helpful to have it working as it had in the past.

This commit rolls the file test_suite/cellblender_test_suite.py back to

  Commit: 3e2eacde1ade0f05d4e3a8c2949eb5b1631e7eb5