Added panel support for an "Execute Once on Export Data Model" script
This is the first step in supporting a data model script that is
executed once for each export (or export and run). This version just
creates the panel controls needed to support the functionality. The
layout is close to what might be needed, but the "refresh" button isn't
right-justified as it is in other places. This may require the use of a
"split" as done in the parameter system (lines 1008-1024). That nicer
formatting can be wrestled with in subsequent versions.