Updated ASCII export label to be more explicit
The previous text for the ASCII export label had been:

   "Export all molecules as ASCII"

However, that doesn't make it clear that BOTH formats are produced.

The new label is:

   "Export ASCII Mol Viz Also"

Which is more descriptive of what is actually being done.

This change was motivated by the discovery of some other bugs in the
exporting of ASCII molecule visualization:

  1. The normal exporter does not export ASCII (but Queue Runner does)
  2. The ASCII files do not seem to be used for visualization at all

Note that there is a check box in the "Visualization Settings" panel
named "Change Viz Data to Ascii", but that doesn't seem to cause the
ASCII data to be read. This can be verified by generating ASCII data
(using the Queue Runner with "Export ASCII Mol Viz Also" enabled),
checking the "Change Viz Data to Ascii", and then deleting the normal
(non-ASCII) files. In that case there is no visualization produced and
an "IOError" is produced on the command line. Further verification can
be done by substituting different mol-viz files in the "viz_data"
directory. Those different files will be shown regardless of the setting
of the "Change Viz Data to Ascii" setting.