Changed string for "ascii_enable" check box to be a bit clearer.
The previous version was:

   "Change Viz Data to Ascii"

The new version is:

   "Generate ASCII Viz Data"

Also changed "Export" to "Generate" for "export_all_ascii" to be
clearer.

Note that the existence of these two flags is confusing because they do
very similar things but are located in different CellBlender panels.

Also note that CellBlender reads whatever data files exist in the
designated directory whether they are binary or ASCII or mixed. Each
file is detected on a file by file basis regardless of the name
("cellbin" or "ascii").