Implemented a more proper fix for the ASCII exporting issue.
As reflected in the previous 2 commits, there has been some confusion
about the ASCII exporting options. This commit improves the situation.

History:

The ASCII export has long been a part of the io_mesh_mcell_mdl exporter
code. The option was triggered in CellBlender in the "Advanced" section
of the Export and Run panel. However, ASCII exporting was not implemented
in the data model exporting code. As a result, the ASCII option worked
properly when run under some of the Engine/Runner pairs but not under
others.

In an attempt to fix this, someone had added the "ascii_enable" Blender
property which duplicated the "export_all_ascii" Blender property (which
was already stored in the Data Model). The newly added "ascii_enable"
property was obtained directly via "mcell.mol_viz.ascii_enable" when
exporting through the "io_mesh_mcell_mdl/export_mcell_mdl.py" file. This
newly added Blender property was NOT stored in the Data Model. To make
things more confusing, the new "ascii_enable" property was added to the
Visualization panel which appeared to duplicate the functionality in the
"Run / Advanced Options" panel, but they actually did different things.

As one more twist, the "export_all_ascii" flag actually enabled ASCII
output in addition to the binary output (producing both ASCII and binary
files in different directories). The binary files continued to be in the
"viz_data" directory, but the ASCII files were placed in a separate
"viz_data_ascii" directory.

This Commit:

To clean this up, the "ascii_enable" Blender property was removed since
it duplicated the intended functionality of the "export_all_ascii"
property. The Visualization panel now displays the "export_all_ascii"
property where it had displayed the "ascii_enable" property. This means
that the "export_all_ascii" flag can be set in 2 places. It can be set
in the Run/Advanced panel (where it had always been) and it can also be
set in the Visualization panel (where the "ascii_enable" had been).
Setting it in either place will update it in the other. The data model
exporter was updated to properly use the "export_all_ascii" flag (from
the data model) to write the proper "MODE" to the MDL ("CELLBLENDER" or
"ASCII"). This appears to work properly at this point, and the flag is
shown in both panels as "ASCII Molecule Files" to make it as clear as
possible what it actually does.

Conclusion:

The "ASCII Molecule Files" option appears in both the "Export and Run"
panel ("Advanced Options") and the "Visualization Settings" panel. It
now works whether exporting direct MDL or exporting through a Data
Model. There is only one molecule visualization directory created when
running and it is named "viz_data" (as it always has been). The
directory named "viz_data_ascii" is no longer generated and the ASCII
files which had been in "viz_data_ascii" are now in "viz_data" when
generating ASCII data. The molecule visualization files are still named
differently according to the previous conventions:

  Scene.cellbin.###.dat   (for binary files)
  Scene.ascii.###.dat     (for ASCII files)

If any code had depended on the existence of the separate "_ascii"
directory, this can be added as a separate option to make it clear
what's being done. Since CellBlender can display from either format,
it's likely that anyone needing the ASCII versions could just use them
from the "viz_data" directory without needing to generate both.