Added changes to support ASCII Exporting in the MCell-R code.
The "write_viz_out" and "write_viz_out_mdlr3" functions both need to
have the "initialization" section of the Data Model passed to them in
order to determine whether to export in ASCII mode or not. The previous
commit did this for the normal "write_viz_out" function, but not for the
"write_viz_out_mdlr3" function. This commit adds that change.