Began adding CellBlender support for a "bending angle" at components.
This commit has begun to add support for a bending angle at components
which would be based on rotating the variable molecule about an axis
through its binding component and normal to the plane defined by the
fixed molecule/component/key points. The angle of rotation about that
axis should bring the binding keys in alignment with when the binding
planes are also aligned with each other.

Unfortunately, there are two axial rotations involved and they are
combined into a single angle (composite_rot_angle) for efficiency. The
first axial rotation is done to align the two key planes. The second
axial rotation is an arbitrary amount specified for this bond. For
efficiency, these two angles are combined (into composite_rot_angle)
and then the rotation matrix is computed from the combined angle. This
is unfortunate because the bending angle should be applied when the two
key planes are aligned and before the arbitrary rotation angle has been
added.