Implemented more proper export of periodic boundary conditions.
In response to GitHub report:
  Defining 'extracellular' molecules #111

The CellBlender example for Periodic Boundary Conditions had been
implemented with an Export Script to generate the "PERIODIC BOX" MDL
required to run the example.

Interestingly, there was a CellBlender Panel for Periodic Boundary
Conditions, but the data was not being used.

This commit adds the code needed to use the Periodic Boundary Conditions
properties when exporting and to also store and retrieve them from the
Data Model. The format for the MDL output was taken from the example.

Note that this commit has not been tested in any cases but the example.