Improved the internal test case in bngl_to_data_model.py for REDEFINE
This version of the internal test case defines these parameters:

   diff_const_multiplier = 1e-10
   MCELL_REDEFINE_diff_const_multiplier = 1

Later, the molecule diffusion constants are defined using that constant:

   MCELL_DEFAULT_DIFFUSION_CONSTANT_2D  1.7e-7 * diff_const_multiplier
   MCELL_DEFAULT_DIFFUSION_CONSTANT_3D  8.51e-7 * diff_const_multiplier
   MCELL_DIFFUSION_CONSTANT_2D_Lyn  1.7e-7 * diff_const_multiplier
   MCELL_DIFFUSION_CONSTANT_2D_Rec  1.7e-7 * diff_const_multiplier
   MCELL_DIFFUSION_CONSTANT_3D_Lig  8.51e-7 * diff_const_multiplier
   MCELL_DIFFUSION_CONSTANT_3D_Syk  8.51e-7 * diff_const_multiplier

When run in BioNetGen, none of those constants are used.

However, when run in MCell, those diffusion constants are applied
to all of the molecules in the model.

If the redefine is working properly, then "diff_const_multiplier"
will be 1 and all of those diffusion constants will be as they
should be for an MCell simulation. They will all move.

However, if the redefine does not work, then "diff_const_multiplier"
will still be 1e-10. This will reduce the final diffusion constants
to nearly zero and none of the molecules will move.

While this is not an exhaustive test of the parser, it will at least
demonstrate that the MCELL_REDEFINE_ is indeed working as planned.
Note that this test can be run by running "bngl_to_data_model.py"
from the command line with the name of an output file. The output
file will be a JSON version of the data model and it can easily
be verified whether the MCELL_REDEFINE_diff_const_multiplier has
actually changed the parameter value of diff_const_multiplier from
1e-10 to 1.