Fix show/hide for model objects
In Blender 2.93 there are two separate settings for object visibility:

The property, object.hide_viewport, controls the global visibility across all
viewports (a property of the Object ID class)

The class methods object.hide_get() and object.hide_set() control the
visibility in the current viewport only (a property of the ObjectBase
bpy_struct class). This property is set/cleared by the Outliner visibilty
("eyeball" icon) control.  There is an option in the outliner to also allow
control of the object.hide_viewport property ("monitor" icon) but this is not
the default setting.

Unless both are set to False (i.e. not hidden) the object will be hidden.

This is cumbersome and confusing and is the origin of problems encountered
during "Export & Run", and when opening old Blend files or importing objects from older Blender files.

Fortunately there is an object class method called object.visible_get() which
makes it easy to check on whether both hide flags are set to False.

This changes forces the two hide flags to be set or cleared together by the
model object panel's visibility controls and during model Export.